backend/requirements.txt
fastapi[all]
supabase
python-dotenv
# google-generativeai
gTTS
google-genai